requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
atproto==0.0.55
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
libipld==3.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
